==> Event Study/3_regression analysis.Rmd <==
---
title: "Regression Analysis for  5d EMBI_Residual"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Ctrl+Alt+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Ctrl+Shift+K* to preview the HTML file).

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed.

```{r}
rm(list = ls())

library(readstata13)
library(tidyr)
library(dplyr)
library(lubridate)

setwd("U:/My Documents/R/Financial Analysis/Event Study")
data_folder <- "C:/Users/PZhao/Box/Effectiveness/Database/2. Database"
saveFolder <- "C:/Users/PZhao/Box/Effectiveness/Database/4. R data"

# Load Main Data with Residuals
load(file = paste(saveFolder, "event_panel_data.Rda", sep ="/"))

```
Baseline on size of BSL relative to GDP
```{r}
fit_0 <- lm(e_embi_5d~size, main_df)
summary(fit_0)
```
Specification 1: dummy=1 if lender is China, zero otherwise, and this variable interacted with Size BSL
```{r}
fit1_1 <- lm(e_embi_5d~size + china, main_df)
summary(fit1_1)
```
```{r}
### add interaction with size
fit1_2 <- lm(e_embi_5d~size * china, main_df)
summary(fit1_2)
```
Specification 2: dummy=1 if BSL is not the first, zero otherwise, and this variable interacted with Size BSL
```{r}
fit2_1 <- lm(e_embi_5d~size+is_first, main_df)
summary(fit2_1)
```
```{r}
###Add interaction
fit2_2 <- lm(e_embi_5d ~size*is_first, main_df)
summary(fit2_2)
```
Specification 3: dummy=1 if BSL is not the first and has been drawn in the past, zero otherwise, and this variable interacted with Size BSL (skipped for now)
Specification 4: whether country is part of RFA  at the time of announcement, and this variable interacted with Size BSL(skipped for now)
Specification 5: previous year's reserves to GDP, and this variable interacted with Size BSL
```{r}
fit5_1 <- lm(e_embi_5d ~size + res_gdp_lag, main_df)
summary(fit5_1)
```
```{r}
#Add interaction
fit5_2 <- lm(e_embi_5d ~size * res_gdp_lag, main_df)
summary(fit5_2)
```
Specification 6: previous year's reserves to M2, and this variable interacted with Size BSL
```{r}
fit6_1 <- lm(e_embi_5d ~size + res_m2_lag, main_df)
summary(fit6_1)
```
```{r}
# Add interaction
fit6_2 <- lm(e_embi_5d ~ size * res_m2_lag, main_df)
summary(fit6_2)
```
Specification 7: previous year's short-term debt due to GDP, and this variable interacted with Size BSL
```{r}
fit7_1 <- lm(e_embi_5d ~size + debt_gdp_lag, main_df)
summary(fit7_1)
```

```{r}
## Add Interaction
fit7_2 <- lm(e_embi_5d ~size * debt_gdp_lag, main_df)
summary(fit7_2)
```
Specification 8: previous year's short-term debt due to M2, and this variable interacted with Size BSL
```{r}
fit8_1 <- lm(e_embi_5d ~size + debt_m2_lag, main_df)
summary(fit8_1)
```
```{r}
# Add interaction
fit8_2 <- lm(e_embi_5d ~size *  debt_m2_lag, main_df)
summary(fit8_2)
```
Specification 9: dummy=1 if overlapping with a fund program , and this variable interacted with Size BSL
```{r}
fit9_1 <- lm(e_embi_5d ~size + program, main_df)
summary(fit9_1)
```
```{r}
# Add Interaction
fit9_2 <- lm(e_embi_5d ~ size * program, main_df)
summary(fit9_2)
```
Specification 10: average spread in the previous month, and this variable interacted with Size BSL
```{r}
fit10_1 <- lm(e_embi_5d ~size + e_embi_l_20d, main_df)
summary(fit10_1)
```
```{r}
# Add residuals
fit10_2 <- lm(e_embi_5d ~size * e_embi_l_20d, main_df)
summary(fit10_2)
```
Specification 11: previous year's current account to GDP , and this variable interacted with Size BSL
